requirements.txt
flask==3.1.1
pillow==11.1.0
ultralytics==8.3.145
pyyaml==6.0.1
labelImg